Fix `Emitter` calculation when using max amount
Fixes #38

+++ confettikit/src/commonMain/kotlin/io/github/vinceglb/confettikit/core/emitter/EmitterConfig.kt
@@ -40,7 +40,7 @@ public class EmitterConfig(
      * Amount of particles created over the duration that is set
      */
     public fun max(amount: Int): EmitterConfig {
-        this.amountPerMs = (emittingTime / amount) / 1000f
+        this.amountPerMs = (emittingTime.toFloat() / amount) / 1000f
         return this
     }
